Update packaging of postgres connector
