+++ packages/apps/package.json
@@ -20,7 +20,7 @@
   },
   "devDependencies": {
     "@counterfactual/contracts": "0.1.1",
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "@types/chai": "4.1.7",
     "@types/mocha": "5.2.7",
     "chai": "4.2.0",
