+++ packages/cf.js/package.json
@@ -30,7 +30,7 @@
   },
   "dependencies": {
     "@counterfactual/node-provider": "0.1.2",
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "ethers": "4.0.28",
     "eventemitter3": "^3.1.2"
   }
