+++ packages/firebase-client/package.json
@@ -12,7 +12,7 @@
     "lint": "tslint -c tslint.json -p ."
   },
   "devDependencies": {
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "@counterfactual/node": "0.1.20",
     "@types/firebase": "3.2.1",
     "rollup": "1.12.3"
