+++ packages/high-roller-bot/package.json
@@ -29,7 +29,7 @@
   },
   "dependencies": {
     "@counterfactual/node": "0.1.20",
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "eventemitter3": "^3.1.2"
   },
   "devDependencies": {
