+++ packages/node-provider/package.json
@@ -17,7 +17,7 @@
     "postinstall": "patch-package"
   },
   "devDependencies": {
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "@types/jest": "24.0.12",
     "@types/node": "12.0.4",
     "jest": "24.8.0",
