+++ packages/playground-server/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "@counterfactual/firebase-client": "0.0.1",
     "@counterfactual/node": "0.1.20",
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "@counterfactual/typescript-typings": "0.1.0",
     "@ebryn/jsonapi-ts": "0.1.17",
     "@koa/cors": "^3.0.0",
