+++ packages/playground/package.json
@@ -14,7 +14,7 @@
   },
   "dependencies": {
     "@counterfactual/node": "0.1.20",
-    "@counterfactual/types": "0.0.9",
+    "@counterfactual/types": "0.0.10",
     "@stencil/core": "0.18.1-0",
     "@stencil/router": "0.3.3",
     "@stencil/sass": "0.2.3",
