+++ packages/postgresql-node-connector/package.json
@@ -1,7 +1,25 @@
 {
   "name": "@counterfactual/postgresql-node-connector",
-  "version": "0.0.1",
+  "version": "0.0.2",
+  "main": "dist/index.js",
   "description": "Implementation of a PostgreSQL storage service for the Counterfactual Node",
   "license": "MIT",
-  "dependencies": {}
+  "scripts": {
+    "build": "tsc -b . && rollup -c"
+  },
+  "dependencies": {
+    "@counterfactual/types": "0.0.10",
+    "pg": "7.11.0",
+    "reflect-metadata": "^0.1.13",
+    "typeorm": "^0.2.17"
+  },
+  "devDependencies": {
+    "@types/node": "12.0.4",
+    "@types/pg": "7.4.14",
+    "rollup": "1.13.0",
+    "rollup-plugin-commonjs": "10.0.0",
+    "rollup-plugin-node-resolve": "5.0.1",
+    "rollup-plugin-typescript2": "0.21.1",
+    "typescript": "3.5.1"
+  }
 }
