+++ packages/postgresql-node-connector/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "declarationDir": "dist/",
+    "outDir": "dist/",
+    "resolveJsonModule": true,
+    "types": ["node"]
+  },
+  "references": [{ "path": "../types" }],
+  "include": ["src", "test"]
+}
