+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@counterfactual/types",
-  "version": "0.0.9",
+  "version": "0.0.10",
   "description": "TypeScript typings for common Counterfactual types",
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
